+++ src/torchmetrics/image/ssim.py
@@ -28,7 +28,7 @@ if not _MATPLOTLIB_AVAILABLE:
 
 
 class StructuralSimilarityIndexMeasure(Metric):
-    """Compute Structual Similarity Index Measure (SSIM_).
+    """Compute Structural Similarity Index Measure (SSIM_).
 
     As input to ``forward`` and ``update`` the metric accepts the following input
 
@@ -230,7 +230,7 @@ class MultiScaleStructuralSimilarityIndexMeasure(Metric):
 
     As output of `forward` and `compute` the metric returns the following output
 
-    - ``msssim`` (: :class:`~torch.Tensor`): if ``reduction!='none'`` returns float scalar tensor with average MSSSIM
+    - ``msssim`` (:class:`~torch.Tensor`): if ``reduction!='none'`` returns float scalar tensor with average MSSSIM
       value over sample else returns tensor of shape ``(N,)`` with SSIM values per sample
 
     Args: