+++ tests/unittests/regression/test_mean_error.py
@@ -132,9 +132,8 @@ def _single_target_ref_metric(preds, target, sk_fn, metric_args):
 def _multi_target_ref_metric(preds, target, sk_fn, metric_args):
     sk_preds = preds.view(-1, num_targets).numpy()
     sk_target = target.view(-1, num_targets).numpy()
-
-    res = sk_fn(sk_target, sk_preds)
-
+    sk_kwargs = {"multioutput": "raw_values"} if metric_args and "num_outputs" in metric_args else {}
+    res = sk_fn(sk_target, sk_preds, **sk_kwargs)
     return math.sqrt(res) if (metric_args and not metric_args["squared"]) else res
 
 
@@ -150,6 +149,7 @@ def _multi_target_ref_metric(preds, target, sk_fn, metric_args):
     [
         (MeanSquaredError, mean_squared_error, sk_mean_squared_error, {"squared": True}),
         (MeanSquaredError, mean_squared_error, sk_mean_squared_error, {"squared": False}),
+        (MeanSquaredError, mean_squared_error, sk_mean_squared_error, {"squared": True, "num_outputs": num_targets}),
         (MeanAbsoluteError, mean_absolute_error, sk_mean_absolute_error, {}),
         (MeanAbsolutePercentageError, mean_absolute_percentage_error, sk_mean_abs_percentage_error, {}),
         (