+++ CHANGELOG.md
@@ -7,6 +7,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 **Note: we move fast, but still we preserve 0.1 version (one feature release) back compatibility.**
 
 
+## [UnReleased] - 2023-08-DD
+
+### Added
+
+-
+
+
+### Changed
+
+-
+
+
+### Fixed
+
+-
+
+
 ## [1.0.2] - 2023-08-02
 
 ### Added